feat: behavior order, description prop and active prop (azion.config.js) (#341)
