+++ lib/presets/next/compute/azion.config.js
@@ -10,29 +10,35 @@ const AzionConfig = {
       {
         name: 'Set Storage Origin for All Requests',
         match: '^\\/_next\\/static\\/', // starts with '/_next/static/'
-        setOrigin: {
-          name: 'origin-storage-default',
-          type: 'object_storage',
+        behavior: {
+          setOrigin: {
+            name: 'origin-storage-default',
+            type: 'object_storage',
+          },
+          deliver: true,
         },
-        deliver: true,
       },
       {
         name: 'Deliver Static Assets',
         match:
           '.(css|js|ttf|woff|woff2|pdf|svg|jpg|jpeg|gif|bmp|png|ico|mp4|json|xml|html)$',
-        setOrigin: {
-          name: 'origin-storage-default',
-          type: 'object_storage',
+        behavior: {
+          setOrigin: {
+            name: 'origin-storage-default',
+            type: 'object_storage',
+          },
+          deliver: true,
         },
-        deliver: true,
       },
       {
         name: 'Execute Edge Function',
         match: '^/',
-        runFunction: {
-          path: '.edge/worker.js',
+        behavior: {
+          runFunction: {
+            path: '.edge/worker.js',
+          },
+          forwardCookies: true,
         },
-        forwardCookies: true,
       },
     ],
   },
