+++ lib/presets/next/deliver/azion.config.js
@@ -10,33 +10,41 @@ const AzionConfig = {
       {
         name: 'Set Storage Origin for All Requests',
         match: '^\\/',
-        setOrigin: {
-          name: 'origin-storage-default',
-          type: 'object_storage',
+        behavior: {
+          setOrigin: {
+            name: 'origin-storage-default',
+            type: 'object_storage',
+          },
         },
       },
       {
         name: 'Deliver Static Assets',
         match:
           '.(css|js|ttf|woff|woff2|pdf|svg|jpg|jpeg|gif|bmp|png|ico|mp4|json|xml|html)$',
-        setOrigin: {
-          name: 'origin-storage-default',
-          type: 'object_storage',
+        behavior: {
+          setOrigin: {
+            name: 'origin-storage-default',
+            type: 'object_storage',
+          },
+          deliver: true,
         },
-        deliver: true,
       },
       {
         name: 'Redirect to index.html',
         match: '.*/$',
-        rewrite: {
-          set: (uri) => `${uri}index.html`,
+        behavior: {
+          rewrite: {
+            set: (uri) => `${uri}index.html`,
+          },
         },
       },
       {
         name: 'Redirect to index.html for Subpaths',
         match: '^(?!.*\\/$)(?![\\s\\S]*\\.[a-zA-Z0-9]+$).*',
-        rewrite: {
-          set: (uri) => `${uri}/index.html`,
+        behavior: {
+          rewrite: {
+            set: (uri) => `${uri}/index.html`,
+          },
         },
       },
     ],
