+++ lib/presets/typescript/compute/azion.config.js
@@ -4,8 +4,10 @@ const AzionConfig = {
       {
         name: 'Execute Edge Function',
         match: '^\\/',
-        runFunction: {
-          path: '.edge/worker.js',
+        behavior: {
+          runFunction: {
+            path: '.edge/worker.js',
+          },
         },
       },
     ],
