+++ lib/presets/vue/deliver/azion.config.js
@@ -10,29 +10,33 @@ const AzionConfig = {
       {
         name: 'Set Storage Origin for All Requests',
         match: '^\\/',
-        setOrigin: {
-          name: 'origin-storage-default',
-          type: 'object_storage',
+        behavior: {
+          setOrigin: {
+            name: 'origin-storage-default',
+            type: 'object_storage',
+          },
         },
       },
-
       {
         name: 'Deliver Static Assets',
         match:
           '.(css|js|ttf|woff|woff2|pdf|svg|jpg|jpeg|gif|bmp|png|ico|mp4|json|xml|html)$',
-        setOrigin: {
-          name: 'origin-storage-default',
-          type: 'object_storage',
+        behavior: {
+          setOrigin: {
+            name: 'origin-storage-default',
+            type: 'object_storage',
+          },
+          deliver: true,
         },
-        deliver: true,
       },
 
       {
         name: 'Redirect to index.html',
         match: '^\\/',
-        // eslint-disable-next-line no-template-curly-in-string
-        rewrite: {
-          set: () => `/index.html`,
+        behavior: {
+          rewrite: {
+            set: () => `/index.html`,
+          },
         },
       },
     ],
