+++ lib/utils/generateManifest/generateManifest.utils.js
@@ -3,7 +3,10 @@ import Ajv from 'ajv';
 import ajvErrors from 'ajv-errors';
 import addKeywords from 'ajv-keywords';
 
-import azionConfigSchema from './fixtures/schema.js';
+import azionConfigSchema from './helpers/schema.js';
+import { requestBehaviors, responseBehaviors } from './helpers/behaviors.js';
+import convertLegacyConfig from './helpers/convertLegacyConfig.js';
+
 /**
  * Validates the provided configuration against a JSON Schema.
  * This function uses AJV (Another JSON Schema Validator) to validate the configuration.
@@ -27,16 +30,18 @@ function validateConfig(config) {
  * Converts the custom configuration to the format expected by the Azion CDN (see https://api.azion.com/).
  * This function processes the custom configuration module to adapt it to the CDN's expected format.
  * It evaluates mathematical expressions, validates types, and transforms configurations for caching and rules.
- * @param {object} config - The configuration object read from the custom configuration module.
+ * @param {object} inputConfig - The configuration object read from the custom configuration module.
  * @returns {object} The converted configuration object suitable for the CDN.
  */
-function jsToJson(config) {
+function jsToJson(inputConfig) {
+  /*  Converts legacy configuration properties to the new `behavior` format. */
+  const config = convertLegacyConfig(inputConfig);
   validateConfig(config);
 
   const payloadCDN = {
     origin: [],
-    rules: [],
     cache: [],
+    rules: [],
   };
 
   // Helper function to safely evaluate mathematical expressions
@@ -101,15 +106,38 @@ function jsToJson(config) {
     });
   }
 
+  // Helper function to add behaviors to a rule
+  const addBehaviors = (cdnRule, behaviors, behaviorDefinitions) => {
+    if (behaviors && typeof behaviors === 'object') {
+      Object.entries(behaviors).forEach(([key, value]) => {
+        if (behaviorDefinitions[key]) {
+          const transformedBehavior = behaviorDefinitions[key].transform(
+            value,
+            payloadCDN,
+          );
+          if (Array.isArray(transformedBehavior)) {
+            cdnRule.behaviors.push(...transformedBehavior);
+          } else if (transformedBehavior) {
+            cdnRule.behaviors.push(transformedBehavior);
+          }
+        } else {
+          console.warn(`Unknown behavior: ${key}`);
+        }
+      });
+    }
+  };
+
   // Convert request rules
-  config?.rules?.request?.forEach((rule) => {
+  config?.rules?.request?.forEach((rule, index) => {
     const cdnRule = {
       name: rule.name,
       phase: 'request',
+      description: rule.description ?? '',
+      is_active: rule.active !== undefined ? rule.active : true, // Default to true if not provided
+      order: index + 2, // index starts at 2, because the default rule is index 1
       criteria: [
         [
           {
-            // eslint-disable-next-line no-template-curly-in-string
             variable: `\${${rule.variable ?? 'uri'}}`,
             operator: 'matches',
             conditional: 'if',
@@ -117,195 +145,23 @@ function jsToJson(config) {
           },
         ],
       ],
-      behaviors: rule.behaviors ? [...rule.behaviors] : [],
+      behaviors: [],
     };
 
-    // Add specific behaviors for the rewrite rule
-    if (rule.rewrite) {
-      const functionString = rule.rewrite.set.toString();
-      const paramNameMatch = functionString.match(/\(([^)]+)\)/);
-      let paramName = paramNameMatch ? paramNameMatch[1] : null;
-      if (paramName) {
-        // eslint-disable-next-line prefer-destructuring
-        paramName = paramName.trim().split(/\s*,\s*/)[0];
-      }
-
-      if (rule.rewrite.match) {
-        cdnRule.behaviors.push({
-          name: 'capture_match_groups',
-          target: {
-            captured_array: paramName || 'captured',
-            // eslint-disable-next-line no-template-curly-in-string
-            variable: `\${${rule.variable ?? 'uri'}}`,
-            regex: rule.rewrite.match,
-          },
-        });
-      }
-      // Transform the set function into a replacement string
-      const pathTransformation = functionString
-        .replace(/.*=>\s*`/, '') // Remove the function part up to the template string
-        .replace(/`\s*;?\s*}?\s*$/, '') // Removes the end of the template string and possible function closures
-        .replace(/\$\{([^}]+)\}/g, (match, p1) => `\${${p1}}`); // Replace ${other[index]} with %{other[index]}
-
-      cdnRule.behaviors.push({
-        name: 'rewrite_request',
-        target: pathTransformation,
-      });
-    }
-
-    // Cache handling within rules
-    if (typeof rule.cache === 'string') {
-      cdnRule.behaviors.push({
-        name: 'set_cache_policy',
-        target: rule.cache,
-      });
-    } else if (typeof rule.cache === 'object') {
-      // Adds the cache object to the cache array
-      const cacheSetting = {
-        name: rule.cache?.name,
-        ...rule.cache, // Here you can expand or adjust fields as necessary
-      };
-      payloadCDN.cache.push(cacheSetting);
-      // Adds the set_cache_policy behavior with the generated cache name
-      cdnRule.behaviors.push({
-        name: 'set_cache_policy',
-        target: rule.cache?.name,
-      });
-      // Updates the rule to use the generated cache name
-      // eslint-disable-next-line no-param-reassign
-      rule.cache = rule.cache?.name;
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'forwardCookies')) {
-      if (rule.forwardCookies) {
-        cdnRule.behaviors.push({
-          name: 'forward_cookies',
-          target: null,
-        });
-      }
-    }
-
-    if (rule.setOrigin) {
-      const origin = payloadCDN.origin.find(
-        (o) =>
-          o.name === rule.setOrigin.name &&
-          o.origin_type === rule.setOrigin.type,
-      );
-
-      if (!origin) {
-        throw new Error(
-          `Rule setOrigin name '${rule.setOrigin.name}' not found in the origin settings`,
-        );
-      } else if (origin.origin_type !== rule.setOrigin.type) {
-        throw new Error(
-          `Rule setOrigin originType '${rule.setOrigin.type}' does not match the origin settings`,
-        );
-      }
-
-      const originBehavior = {
-        name: 'set_origin',
-        target: origin.name,
-      };
-
-      cdnRule.behaviors.push(originBehavior);
-    }
-
-    if (rule.runFunction) {
-      const runFunctionBehavior = {
-        name: 'run_function',
-        target: rule.runFunction.path,
-      };
-
-      cdnRule.behaviors.push(runFunctionBehavior);
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'deliver')) {
-      if (rule.deliver) {
-        cdnRule.behaviors.push({
-          name: 'deliver',
-          target: null,
-        });
-      }
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'enableGZIP')) {
-      if (rule.enableGZIP) {
-        cdnRule.behaviors.push({
-          name: 'enable_gzip',
-          target: '',
-        });
-      }
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'bypassCache')) {
-      if (rule.bypassCache) {
-        cdnRule.behaviors.push({
-          name: 'bypass_cache_phase',
-          target: null,
-        });
-      }
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'httpToHttps')) {
-      if (rule.httpToHttps) {
-        cdnRule.behaviors.push({
-          name: 'redirect_http_to_https',
-          target: null,
-        });
-      }
-    }
-
-    if (rule.redirectTo301) {
-      cdnRule.behaviors.push({
-        name: 'redirect_to_301',
-        target: rule.redirectTo301,
-      });
-    }
-
-    if (rule.redirectTo302) {
-      cdnRule.behaviors.push({
-        name: 'redirect_to_302',
-        target: rule.redirectTo302,
-      });
-    }
-
-    if (rule.setCookie) {
-      const cookieBehavior = {
-        name: 'add_request_cookie',
-        target: rule.setCookie,
-      };
-      cdnRule.behaviors.push(cookieBehavior);
-    }
-
-    if (rule.setHeaders) {
-      rule.setHeaders.forEach((header) => {
-        cdnRule.behaviors.push({
-          name: 'add_request_header',
-          target: header,
-        });
-      });
-    }
-
-    if (rule.capture) {
-      cdnRule.behaviors.push({
-        name: 'capture_match_groups',
-        target: {
-          regex: rule.capture.match,
-          captured_array: rule.capture.captured,
-          subject: `\${${rule.capture.subject ?? 'uri'}}`,
-        },
-      });
-    }
+    addBehaviors(cdnRule, rule.behavior, requestBehaviors);
 
     payloadCDN.rules.push(cdnRule);
   });
 
-  // response rules
+  // Convert response rules
   if (config?.rules?.response) {
-    config.rules.response.forEach((rule) => {
+    config.rules.response.forEach((rule, index) => {
       const cdnRule = {
         name: rule.name,
         phase: 'response',
+        description: rule.description ?? '',
+        is_active: rule.active !== undefined ? rule.active : true, // Default to true if not provided
+        order: index + 2, // index starts at 2, because the default rule is index 1
         criteria: [
           [
             {
@@ -316,88 +172,10 @@ function jsToJson(config) {
             },
           ],
         ],
-        behaviors: rule.behaviors ? [...rule.behaviors] : [],
+        behaviors: [],
       };
 
-      if (rule.setCookie) {
-        cdnRule.behaviors.push({
-          name: 'add_response_cookie',
-          target: rule.setCookie,
-        });
-      }
-
-      if (rule.setHeaders) {
-        rule.setHeaders.forEach((header) => {
-          cdnRule.behaviors.push({
-            name: 'add_response_header',
-            target: header,
-          });
-        });
-      }
-
-      if (Object.prototype.hasOwnProperty.call(rule, 'enableGZIP')) {
-        if (rule.enableGZIP) {
-          cdnRule.behaviors.push({
-            name: 'enable_gzip',
-            target: '',
-          });
-        }
-      }
-
-      if (rule.filterCookie) {
-        cdnRule.behaviors.push({
-          name: 'filter_response_cookie',
-          target: rule.filterCookie,
-        });
-      }
-
-      if (rule.filterHeader) {
-        cdnRule.behaviors.push({
-          name: 'filter_response_header',
-          target: rule.filterHeader,
-        });
-      }
-
-      if (Object.prototype.hasOwnProperty.call(rule, 'deliver')) {
-        if (rule.deliver) {
-          cdnRule.behaviors.push({
-            name: 'deliver',
-            target: null,
-          });
-        }
-      }
-
-      if (rule.runFunction) {
-        cdnRule.behaviors.push({
-          name: 'run_function',
-          target: rule.runFunction.path,
-        });
-      }
-
-      if (rule.redirectTo301) {
-        cdnRule.behaviors.push({
-          name: 'redirect_to_301',
-          target: rule.redirectTo301,
-        });
-      }
-
-      if (rule.redirectTo302) {
-        cdnRule.behaviors.push({
-          name: 'redirect_to_302',
-          target: rule.redirectTo302,
-        });
-      }
-
-      if (rule.capture) {
-        cdnRule.behaviors.push({
-          name: 'capture_match_groups',
-          target: {
-            regex: rule.capture.match,
-            captured_array: rule.capture.captured,
-            subject: `\${${rule.capture.subject ?? 'uri'}}`,
-          },
-        });
-      }
+      addBehaviors(cdnRule, rule.behavior, responseBehaviors);
 
       payloadCDN.rules.push(cdnRule);
     });
