+++ lib/utils/generateManifest/generateManifest.utils.test.js
@@ -29,10 +29,12 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/test',
-              cache: {
-                name: 'directCache',
-                browser_cache_settings_maximum_ttl: 300,
-                cdn_cache_settings_maximum_ttl: 600,
+              behavior: {
+                setCache: {
+                  name: 'directCache',
+                  browser_cache_settings_maximum_ttl: 300,
+                  cdn_cache_settings_maximum_ttl: 600,
+                },
               },
             },
           ],
@@ -55,14 +57,16 @@ describe('Utils - generateManifest', () => {
             {
               name: 'rewriteRule',
               match: '/path',
-              capture: {
-                match: '^(./)([^/])$',
-                captured: 'other',
-                subject: 'uri',
-              },
-              rewrite: {
-                match: '^(./)([^/])$',
-                set: (other) => `/${other[0]}/${other[1]}`,
+              behavior: {
+                capture: {
+                  match: '^(./)([^/])$',
+                  captured: 'other',
+                  subject: 'uri',
+                },
+                rewrite: {
+                  match: '^(./)([^/])$',
+                  set: (other) => `/${other[0]}/${other[1]}`,
+                },
               },
             },
           ],
@@ -84,7 +88,33 @@ describe('Utils - generateManifest', () => {
           expect.objectContaining({
             name: 'rewrite_request',
             // eslint-disable-next-line no-template-curly-in-string
-            target: '/${other[0]}/${other[1]}',
+            target: '/%{other[0]}/%{other[1]}',
+          }),
+        ]),
+      );
+    });
+
+    it('should handle rewrites directly as a string', () => {
+      const azionConfig = {
+        rules: {
+          request: [
+            {
+              name: 'simpleRewriteRule',
+              match: '/simple',
+              behavior: {
+                rewrite: '/new-path',
+              },
+            },
+          ],
+        },
+      };
+
+      const result = jsToJson(azionConfig);
+      expect(result.rules[0].behaviors).toEqual(
+        expect.arrayContaining([
+          expect.objectContaining({
+            name: 'rewrite_request',
+            target: '/new-path',
           }),
         ]),
       );
@@ -120,8 +150,10 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/no-cache',
-              runFunction: {
-                path: '.edge/worker.js',
+              behavior: {
+                runFunction: {
+                  path: '.edge/worker.js',
+                },
               },
             },
           ],
@@ -147,9 +179,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -224,7 +258,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              forwardCookies: true,
+              behavior: {
+                forwardCookies: true,
+              },
             },
           ],
         },
@@ -249,7 +285,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/no-forward',
-              forwardCookies: false,
+              behavior: {
+                forwardCookies: false,
+              },
             },
             {
               name: 'testRule',
@@ -277,6 +315,7 @@ describe('Utils - generateManifest', () => {
         ]),
       );
     });
+
     it('should correctly process the setOrigin rule with all optional fields provided', () => {
       const azionConfigWithAllFields = {
         origin: [
@@ -292,9 +331,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/_next',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -318,9 +359,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/_next_no_type',
-              setOrigin: {
-                name: 'nameWithoutType',
-                // type is missing
+              behavior: {
+                setOrigin: {
+                  name: 'nameWithoutType',
+                  // type is missing
+                },
               },
             },
           ],
@@ -354,8 +397,10 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/run-function-test-path-only',
-              runFunction: {
-                path: '.edge/worker.js',
+              behavior: {
+                runFunction: {
+                  path: '.edge/worker.js',
+                },
               },
             },
           ],
@@ -373,20 +418,23 @@ describe('Utils - generateManifest', () => {
         ]),
       );
     });
+
     it('should include the behavior deliver when deliver is true', () => {
-      const azionConfigComDeliver = {
+      const azionConfig = {
         rules: {
           request: [
             {
               name: 'testRule',
               match: '/path',
-              deliver: true,
+              behavior: {
+                deliver: true,
+              },
             },
           ],
         },
       };
 
-      const result = jsToJson(azionConfigComDeliver);
+      const result = jsToJson(azionConfig);
       expect(result.rules[0].behaviors).toEqual(
         expect.arrayContaining([
           expect.objectContaining({
@@ -395,6 +443,7 @@ describe('Utils - generateManifest', () => {
         ]),
       );
     });
+
     it('should throw an error if deliver is not a boolean', () => {
       const azionConfig = {
         rules: {
@@ -402,7 +451,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/path',
-              deliver: 'true', // Incorretamente definido como string
+              behavior: {
+                deliver: 'true', // Incorrectly defined as string
+              },
             },
           ],
         },
@@ -420,7 +471,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              setCookie: true, // Incorretamente definido como boolean
+              behavior: {
+                setCookie: true, //  Incorrectly defined as boolean
+              },
             },
           ],
         },
@@ -438,7 +491,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              setHeaders: { key: 'value' }, // Incorretamente definido como objeto
+              behavior: {
+                setHeaders: { key: 'value' }, // Incorrectly defined as object
+              },
             },
           ],
         },
@@ -456,7 +511,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/path',
-              deliver: true,
+              behavior: {
+                deliver: true,
+              },
             },
           ],
         },
@@ -479,7 +536,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              setCookie: 'sessionId=abc123',
+              behavior: {
+                setCookie: 'sessionId=abc123',
+              },
             },
           ],
         },
@@ -503,7 +562,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              setHeaders: ['Authorization: Bearer abc123'], // Corretamente definido como uma array de strings
+              behavior: {
+                setHeaders: ['Authorization: Bearer abc123'], // Corretamente definido como uma array de strings
+              },
             },
           ],
         },
@@ -537,9 +598,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -575,9 +638,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'another origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'another origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -604,9 +669,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'another_type',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'another_type',
+                },
               },
             },
           ],
@@ -625,9 +692,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -655,9 +724,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'another_type',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'another_type',
+                },
               },
             },
           ],
@@ -684,9 +755,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my single origin',
-                type: 'single_origin',
+              behavior: {
+                setOrigin: {
+                  name: 'my single origin',
+                  type: 'single_origin',
+                },
               },
             },
           ],
@@ -718,6 +791,7 @@ describe('Utils - generateManifest', () => {
       );
     });
   });
+
   it('should correctly handle bypassCache behavior', () => {
     const azionConfig = {
       rules: {
@@ -725,7 +799,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testBypassCache',
             match: '/',
-            bypassCache: true,
+            behavior: {
+              bypassCache: true,
+            },
           },
         ],
       },
@@ -741,6 +817,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly handle redirect to 301', () => {
     const azionConfig = {
       rules: {
@@ -748,7 +825,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testRedirect301',
             match: '/',
-            redirectTo301: 'https://example.com',
+            behavior: {
+              redirectTo301: 'https://example.com',
+            },
           },
         ],
       },
@@ -772,7 +851,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testRedirect302',
             match: '/',
-            redirectTo302: 'https://example.com',
+            behavior: {
+              redirectTo302: 'https://example.com',
+            },
           },
         ],
       },
@@ -788,6 +869,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly handle capture match groups', () => {
     const azionConfig = {
       rules: {
@@ -795,10 +877,12 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testCapture',
             match: '/',
-            capture: {
-              match: '^/user/(.*)',
-              captured: 'userId',
-              subject: 'uri',
+            behavior: {
+              capture: {
+                match: '^/user/(.*)',
+                captured: 'userId',
+                subject: 'uri',
+              },
             },
           },
         ],
@@ -820,6 +904,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly handle filterCookie behavior', () => {
     const azionConfig = {
       rules: {
@@ -827,7 +912,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testFilterCookie',
             match: '/',
-            filterCookie: '_cookie',
+            behavior: {
+              filterCookie: '_cookie',
+            },
           },
         ],
       },
@@ -843,6 +930,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly process rules in the response phase', () => {
     const azionConfig = {
       rules: {
@@ -850,7 +938,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testResponsePhase',
             match: '/',
-            setHeaders: ['X-Test-Header: value'],
+            behavior: {
+              setHeaders: ['X-Test-Header: value'],
+            },
           },
         ],
       },
@@ -866,6 +956,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly add multiple response headers', () => {
     const azionConfig = {
       rules: {
@@ -873,10 +964,12 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testMultipleHeaders',
             match: '/',
-            setHeaders: [
-              'X-Frame-Options: DENY',
-              "Content-Security-Policy: default-src 'self'",
-            ],
+            behavior: {
+              setHeaders: [
+                'X-Frame-Options: DENY',
+                "Content-Security-Policy: default-src 'self'",
+              ],
+            },
           },
         ],
       },
@@ -896,6 +989,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly handle enableGZIP behavior', () => {
     const azionConfig = {
       rules: {
@@ -903,6 +997,226 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testEnableGZIP',
             match: '/',
+            behavior: {
+              enableGZIP: true,
+            },
+          },
+        ],
+      },
+    };
+
+    const result = jsToJson(azionConfig);
+    expect(result.rules[0].behaviors).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          name: 'enable_gzip',
+          target: '',
+        }),
+      ]),
+    );
+  });
+
+  it('should handle rules with description and active properties correctly', () => {
+    const azionConfig = {
+      rules: {
+        request: [
+          {
+            name: 'Example Rule',
+            match: '/',
+            description: 'This rule redirects all traffic.',
+            active: false,
+          },
+          {
+            name: 'Second Rule',
+            match: '/api',
+            behavior: {},
+            // description is not provided here
+            active: true,
+          },
+          {
+            name: 'Third Rule',
+            match: '/home',
+            description: 'This rule handles home traffic.',
+            behavior: {},
+            // active is not provided here
+          },
+        ],
+      },
+    };
+
+    const result = jsToJson(azionConfig);
+    expect(result.rules).toEqual([
+      expect.objectContaining({
+        name: 'Example Rule',
+        description: 'This rule redirects all traffic.',
+        is_active: false,
+      }),
+      expect.objectContaining({
+        name: 'Second Rule',
+        description: '', // Should default to an empty string
+        is_active: true,
+      }),
+      expect.objectContaining({
+        name: 'Third Rule',
+        description: 'This rule handles home traffic.',
+        is_active: true, // Should default to true
+      }),
+    ]);
+  });
+
+  it('should correctly assign order starting from 2 for request and response rules', () => {
+    const azionConfig = {
+      rules: {
+        request: [
+          { name: 'First Request Rule', match: '/', behavior: {} },
+          { name: 'Second Request Rule', match: '/second', behavior: {} },
+        ],
+        response: [
+          { name: 'First Response Rule', match: '/', behavior: {} },
+          { name: 'Second Response Rule', match: '/second', behavior: {} },
+        ],
+      },
+    };
+
+    const result = jsToJson(azionConfig);
+    expect(result.rules[0]).toEqual(
+      expect.objectContaining({
+        name: 'First Request Rule',
+        order: 2,
+      }),
+    );
+    expect(result.rules[1]).toEqual(
+      expect.objectContaining({
+        name: 'Second Request Rule',
+        order: 3,
+      }),
+    );
+    expect(result.rules[2]).toEqual(
+      expect.objectContaining({
+        name: 'First Response Rule',
+        order: 2,
+      }),
+    );
+    expect(result.rules[3]).toEqual(
+      expect.objectContaining({
+        name: 'Second Response Rule',
+        order: 3,
+      }),
+    );
+  });
+
+  it('should maintain the order of behaviors as specified by the user', () => {
+    const azionConfig = {
+      origin: [
+        {
+          name: 'my origin storage',
+          type: 'object_storage',
+          bucket: 'mybucket',
+          prefix: 'myfolder',
+        },
+      ],
+      rules: {
+        request: [
+          {
+            name: 'testRule',
+            match: '/',
+            behavior: {
+              setHeaders: ['Authorization: Bearer abc123'],
+              deliver: true,
+              setOrigin: {
+                name: 'my origin storage',
+                type: 'object_storage',
+              },
+            },
+          },
+        ],
+      },
+    };
+
+    const result = jsToJson(azionConfig);
+    expect(result.rules[0].behaviors).toEqual([
+      expect.objectContaining({
+        name: 'add_request_header',
+        target: 'Authorization: Bearer abc123',
+      }),
+      expect.objectContaining({
+        name: 'deliver',
+      }),
+      expect.objectContaining({
+        name: 'set_origin',
+        target: 'my origin storage',
+      }),
+    ]);
+  });
+  it('should throw an error when the origin settings are not defined', () => {
+    const azionConfigWithoutOrigin = {
+      rules: {
+        request: [
+          {
+            name: 'testRule',
+            match: '/api',
+            behavior: {
+              setOrigin: {
+                name: 'undefined origin',
+                type: 'object_storage',
+              },
+            },
+          },
+        ],
+      },
+    };
+
+    expect(() => jsToJson(azionConfigWithoutOrigin)).toThrow(
+      "Rule setOrigin name 'undefined origin' not found in the origin settings",
+    );
+  });
+  it('should handle legacy config without behavior field for request rules', () => {
+    const azionConfig = {
+      origin: [
+        {
+          name: 'legacy origin',
+          type: 'object_storage',
+          addresses: ['teste.com'],
+        },
+      ],
+      rules: {
+        request: [
+          {
+            name: 'legacyRule',
+            match: '/legacy',
+            setOrigin: {
+              name: 'legacy origin',
+              type: 'object_storage',
+            },
+            setHeaders: ['Authorization: Bearer legacy'],
+          },
+        ],
+      },
+    };
+
+    const result = jsToJson(azionConfig);
+    expect(result.rules[0].behaviors).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          name: 'set_origin',
+          target: 'legacy origin',
+        }),
+        expect.objectContaining({
+          name: 'add_request_header',
+          target: 'Authorization: Bearer legacy',
+        }),
+      ]),
+    );
+  });
+
+  it('should handle legacy config without behavior field for response rules', () => {
+    const azionConfig = {
+      rules: {
+        response: [
+          {
+            name: 'legacyResponseRule',
+            match: '/legacy-response',
+            setHeaders: ['X-Legacy-Header: legacy'],
             enableGZIP: true,
           },
         ],
@@ -912,6 +1226,82 @@ describe('Utils - generateManifest', () => {
     const result = jsToJson(azionConfig);
     expect(result.rules[0].behaviors).toEqual(
       expect.arrayContaining([
+        expect.objectContaining({
+          name: 'add_response_header',
+          target: 'X-Legacy-Header: legacy',
+        }),
+        expect.objectContaining({
+          name: 'enable_gzip',
+          target: '',
+        }),
+      ]),
+    );
+  });
+
+  it('should handle mixed legacy and new config for request rules', () => {
+    const azionConfig = {
+      origin: [
+        {
+          name: 'mixed origin',
+          type: 'object_storage',
+          addresses: ['teste.com'],
+        },
+      ],
+      rules: {
+        request: [
+          {
+            name: 'mixedRule',
+            match: '/mixed',
+            setOrigin: {
+              name: 'mixed origin',
+              type: 'object_storage',
+            },
+            behavior: {
+              setHeaders: ['Authorization: Bearer mixed'],
+            },
+          },
+        ],
+      },
+    };
+
+    const result = jsToJson(azionConfig);
+    expect(result.rules[0].behaviors).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          name: 'set_origin',
+          target: 'mixed origin',
+        }),
+        expect.objectContaining({
+          name: 'add_request_header',
+          target: 'Authorization: Bearer mixed',
+        }),
+      ]),
+    );
+  });
+
+  it('should handle mixed legacy and new config for response rules', () => {
+    const azionConfig = {
+      rules: {
+        response: [
+          {
+            name: 'mixedResponseRule',
+            match: '/mixed-response',
+            setHeaders: ['X-Mixed-Header: mixed'],
+            behavior: {
+              enableGZIP: true,
+            },
+          },
+        ],
+      },
+    };
+
+    const result = jsToJson(azionConfig);
+    expect(result.rules[0].behaviors).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          name: 'add_response_header',
+          target: 'X-Mixed-Header: mixed',
+        }),
         expect.objectContaining({
           name: 'enable_gzip',
           target: '',
